web/requirements.txt
beautifulsoup4==4.10.0
Django==4.0.3
django-apscheduler==0.6.2
djangorestframework==3.13
python-telegram-bot==13.11
requests==2.27.1
uWSGI==2.0.20
lxml==4.7.1
psycopg2==2.8.6
